BT : Fix UBSan errors in BT Stack
Fix:
This Patch will fix UBSan errors in BT Stack

Test:
BT Basic Sanity

Bug: 161659385

+++ system/stack/l2cap/l2c_utils.cc
@@ -2857,7 +2857,7 @@ void l2cu_process_fixed_disc_cback(tL2C_LCB* p_lcb) {
             xx + L2CAP_FIRST_FIXED_CHNL, p_lcb->remote_bd_addr, false,
             p_lcb->DisconnectReason(), p_lcb->transport);
       }
-    } else if ((peer_channel_mask & (1 << (xx + L2CAP_FIRST_FIXED_CHNL))) &&
+    } else if ((peer_channel_mask & ((uint64_t)1 << (xx + L2CAP_FIRST_FIXED_CHNL))) &&
                (l2cb.fixed_reg[xx].pL2CA_FixedConn_Cb != NULL))
       (*l2cb.fixed_reg[xx].pL2CA_FixedConn_Cb)(
           xx + L2CAP_FIRST_FIXED_CHNL, p_lcb->remote_bd_addr, false,
